requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.6.2
charset-normalizer==3.3.2
Django==4.2.13
googlesearch-python==1.2.4
idna==3.7
requests==2.32.3
soupsieve==2.5
sqlparse==0.5.0
urllib3==2.2.2
